Add files via upload
Brief Description:  
These programs are used to calculate ion scattering parameters such as cross sections, kinematic factors, energies of reaction products, charge states of ions passing through foils, and optimum settings of accelerators to obtain beams.

The IBA Periodic Table.pptx PowerPoint file presents a plot of the table of all elements with the sensitivity, depth resolution and range of 5 IBA techniques.  XSs.xls gives a snapshot of all the IBA cross sections for all the techniques and ion species and energies typically used.  This program was the basis for the sensitivities listed in the IBA Periodic Table.  The elec nuc SP and range vs M and E.ppt file plots the electronic and nuclear stopping powers and ranges for all ions (using interpolation) at all energies in Silicon.

There are 5 MS Excel programs in this “Suite”.  For all of these programs, equations of various complexities are used in the calculations ranging from simple expressions to numerical integrals performed using VBA Excel macros.  The independent variables are input in colored cells of the spreadsheets by the user, and the dependent variables are then calculated based on those inputs and output on the same sheet.

DEDX6.xls uses the equations found in the three Handbooks listed in the References to calculate the electronic and nuclear stopping powers of all ions in the periodic table, accelerated to energies up to 1 GeV, in all material compounds containing up to 6 elements.  The ZBL theory is used for the electronic stopping, while the LSS theory is used for the nuclear stopping.  All of the details of these theories can be found in the Handbooks.

IBA.xls calculates the energy of nuclear scattering or reaction products in the form:
A + B => C + D + Q, where A is the incident particle, B is the target nucleus, C and D are the scattering and/or nuclear reaction products, and Q is the Energy “Q-value” of the reaction.  The rest masses of A-D and the Q value are determined from a lookup table based on the Chart of the Nuclides.  The user enters the independent parameters of the atomic number, weight and energy of A (Z,A) for A-D, and the incident energy of A (E) , and the angle at which the reaction product is detected, and the dependent parameters are the reaction product’s energy and reaction cross section are calculated by the program, again based on the kinematic and elastic scattering equations to be found in the three IBA Handbooks.

Q.xls calculates the distribution of ion’s charges after passage through either low pressures of gas or thin carbon foils.  The independent parameters are the ions Z, A and Energy, and the calculated dependent parameters are the fraction of ions in each of their possible charge states.  The equations given in Ref. 4 below are used in this calculation.

Q-optimum.xls builds off of the Q.xls program above by calculating the charge state and terminal voltage to select when accelerating ions with a Tandem Van de Graaff accelerator that uses gas stripping at the terminal.  The input parameters here are Z, A and Energy, and the dependent calculated parameters are the ion’s charge state and the terminal voltage to use that provides the most intense beam.


Deflection by earth B field.xlsm:  calculates the deflection an accelerated particle beam experiences due to passage through the magnetic field of the earth. 
 
Magnetic deflection of ions.xlsx:  calculates the radius of curvature and deflection angle of accelerated ions through magnetic fields provided by circular pole electromagnets.

Electrostatic deflection of ions.xls:  calculates the deflection angle for accelerated ions passing through a constant electrostatic field.

Channeling.xlsm:  calculates the half-angles for both axial and planar channeling for indexes up to <555> and [555].